server/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4